refactor: promote analyze.py to analyze/ package (#70)
Move existing analyze logic to swebench/analyze/summary.py, re-export
analyze_command as a click.group from swebench/analyze/__init__.py.
No functional change — byte-identical output on golden fixtures.

Add tests/test_cli_smoke.py covering --help for add/run/analyze/cache.
Update tests/test_analyze_summary.py to invoke `analyze summary` (the
verb moved one level deeper).

After pulling: find . -name __pycache__ -delete
