+++ swebench/analyze/__init__.py
@@ -1,0 +1,17 @@
+"""`analyze` command group — log analysis pipeline subcommands."""
+
+from __future__ import annotations
+
+import click
+
+from swebench.analyze.summary import _register
+
+
+@click.group("analyze")
+def analyze_command() -> None:
+    """Analyze SWE-bench results (summary table, etc.)."""
+
+
+_register(analyze_command)
+
+__all__ = ["analyze_command"]
